Update patch set 5
Patch Set 5:

I haven't reviewed the patch completely and you might already have plans to do what I am about to ask so please take the comment as a suggestion.

It is a common practice for majority of the OPNFV projects to put the Jenkins facing scripts in to ci folder in project repo so Jenkins jobs only care about and execute those scripts. The scripts in ci folder can simply be wrappers to the real scripts that do the stuff that matters by passing arguments on so on.

With this, you can do your work in a bit isolated way without touching the scripts located in ci folder - if the work doesn't require changes in these scripts obviously.

Patch-set: 5